(feat) Server i typescript. Jest-config
