+++ .gitignore
@@ -3,3 +3,4 @@ build
 .idea
 .env
 .eslintcache
+node_dist/
