+++ package.json
@@ -67,13 +67,28 @@
     "request": "^2.88.2"
   },
   "scripts": {
-    "format-all": "prettier --write \"src/**/*.{js,jsx,ts,tsx}\" *.js",
+    "format-all": "prettier --write \"src/**/*.{js,jsx,ts,tsx}\"",
     "build_and_serve": "yarn build; yarn serve",
-    "start": "craco start",
-    "build": "craco build",
+    "dev": "node --loader ts-node/esm --experimental-specifier-resolution=node src/server.ts",
+    "build": "craco build && tsc -p tsconfig.backend.json",
     "test-ci": "CI=true craco build",
-    "test": "craco test",
-    "serve": "NODE_ENV=production node server.js"
+    "test": "jest",
+    "serve": "NODE_ENV=production node node_dist/server.js"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "moduleNameMapper": {
+      "\\.(css|less|svg)$": "<rootDir>/src/mockStyle.js",
+      "nav-(.*)-style": "<rootDir>/src/mockStyle.js"
+    },
+    "globals": {
+      "ts-jest": {
+        "tsconfig": "<rootDir>/tsconfig.json"
+      }
+    },
+    "setupFilesAfterEnv": [
+      "<rootDir>/src/jest-setup.ts"
+    ]
   },
   "husky": {
     "hooks": {
@@ -100,13 +115,19 @@
   },
   "devDependencies": {
     "@craco/craco": "^6.0.0",
+    "@types/express": "^4.17.11",
+    "@types/express-session": "^1.17.3",
+    "@types/mustache-express": "^1.2.1",
     "craco-less": "^1.17.0",
     "eslint-config-prettier": "^7.2.0",
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-react-app": "^6.2.2",
     "http-proxy-middleware": "^1.0.5",
     "husky": "^4.3.8",
+    "identity-obj-proxy": "^3.0.0",
     "prettier": "^2.0.5",
+    "ts-jest": "^26.4.4",
+    "ts-node": "^9.1.1",
     "typescript": "^4.1.3"
   }
 }
