+++ src/server.ts
@@ -4,20 +4,21 @@ import dotenv from 'dotenv';
 import express from 'express';
 import mustacheExpress from 'mustache-express';
 
-import getDecorator from './src/dekorator';
-import setupProxy from './src/setupProxy';
+import getDecorator from './dekorator';
+import { createApiForwardingFunction } from './proxy';
+
 dotenv.config();
 
 const app = express();
 
-app.set('views', `${__dirname}/build`);
+app.set('views', `${__dirname}/../build`);
 app.set('view engine', 'mustache');
 app.engine('html', mustacheExpress());
 
-setupProxy(app);
+app.use('/api', createApiForwardingFunction());
 
 // Static files
-app.use(express.static(path.join(__dirname, 'build'), { index: false }));
+app.use(express.static(path.join(__dirname, '..', 'build'), { index: false }));
 
 // Nais functions
 app.get(`/internal/isAlive|isReady`, (_req, res) => res.sendStatus(200));
@@ -25,6 +26,7 @@ app.get(`/internal/isAlive|isReady`, (_req, res) => res.sendStatus(200));
 app.get('*', (_req, res) =>
     getDecorator()
         .then(fragments => {
+            // @ts-ignore
             res.render('index.html', fragments);
         })
         .catch(e => {
