+++ tsconfig.json
@@ -1,26 +1,27 @@
 {
+  "extends": "./tsconfig.base.json",
   "compilerOptions": {
-    "target": "es6",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
+    "outDir": "node_dist",
+    "baseUrl": "./src",
+    "rootDir": "./src",
+    "resolveJsonModule": true,
     "allowJs": true,
     "skipLibCheck": true,
-    "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
-    "strict": true,
     "forceConsistentCasingInFileNames": true,
     "module": "esnext",
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react-jsx",
-    "noFallthroughCasesInSwitch": true
+    "jsx": "react-jsx"
   },
   "include": [
-    "src"
+    "./src"
+  ],
+  "exclude": [
+    "(.*).test.(.*)"
+  ],
+  "lib": [
+    "es2015",
+    "esnext"
   ]
 }
